修改 replyStatusID 从 int类型转换成long类型

+++ server/src/weibo4j/Comment.java
@@ -25,7 +25,7 @@ public class Comment extends WeiboResponse implements java.io.Serializable {
     private String source;
     private boolean isTruncated;
     private long inReplyToStatusId;
-    private int inReplyToUserId;
+    private long inReplyToUserId;
     private boolean isFavorited;
     private String inReplyToScreenName;
     private double latitude = -1;
@@ -204,7 +204,7 @@ public class Comment extends WeiboResponse implements java.io.Serializable {
      * @return the in_reply_tostatus_id
      * @since Weibo4J 1.0.4
      */
-    public int getInReplyToUserId() {
+    public long getInReplyToUserId() {
         return inReplyToUserId;
     }
